[payment] add payment

+++ db/migrate/20240104110414_create_cart_items.rb
@@ -4,7 +4,7 @@ class CreateCartItems < ActiveRecord::Migration[7.1]
       t.references :cart, null: false, foreign_key: true
       t.references :event_seat, null: false, foreign_key: true
       t.integer :price_estimate_rule, null: false, default: 0
-      t.float :price, null: false, default: 0
+      t.integer :price, null: false, default: 0
 
       t.timestamps
     end
