+++ db/migrate/20240108064635_create_orders.rb
@@ -3,7 +3,7 @@ class CreateOrders < ActiveRecord::Migration[7.1]
     create_table :orders do |t|
       t.references :cart, null: false, foreign_key: true
       t.integer :status, null: false, default: 0
-      t.float :total_price, null: false, default: 0
+      t.integer :total_price, null: false, default: 0
       t.string :description
 
       t.timestamps
